% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMediarithmics.R
\name{loadMediarithmics.ad}
\alias{loadMediarithmics.ad}
\title{loadMediarithmics}
\usage{
loadMediarithmics.ad(client, datadir, dfs.names)
}
\arguments{
\item{client}{}

\item{datadir}{the directory of the data}

\item{dfs.names}{}
}
\value{
A data.frame with header as given in input
}
\description{
loadMediarithmics
}

