% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWeborama.R
\name{loadWeborama.ad}
\alias{loadWeborama.ad}
\title{loadWeborama.Ad}
\usage{
loadWeborama.ad(client, datadir, dfs.names)
}
\arguments{
\item{client}{}

\item{datadir}{the directory of the data}

\item{dfs.names}{}
}
\value{
An object from dplyr with colnames given by the input \code{dfs.names}
}
\description{
A function for loading data exported from the Weborama interface
}

