% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{makeReport}
\usage{
makeReport(client, dfs.names, ...)
}
\arguments{
\item{client}{}

\item{dfs.names}{the header of the data}

\item{...}{other parameters given to writeExcel}
}
\description{
This function creates the report based on the data found in the folders Weborama and Mediarithmics
}

