% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeExcel.R
\name{writeExcel}
\alias{writeExcel}
\title{writeExcel}
\usage{
writeExcel(webo.ad, media.ad, webo.goals, media.goals, budget, marge,
  dfs.names = names(webo.ad), save.dir)
}
\arguments{
\item{webo.ad}{the data.frame for ad data from Weborama}

\item{media.ad}{the data.frame for ad data from Mediarithmics}

\item{webo.goals}{the data.frame for goals data from Weborama}

\item{media.goals}{the data.frame for goals data from Mediarithmics}

\item{budget}{the total budget to normalise the spent}

\item{marge}{the coefficient to mutliply the spent by}

\item{dfs.names}{the header of the excel file}

\item{save.dir}{the saving directory with full name, ie <directory>/<filename>.xlsx}
}
\description{
writeExcel
}

